it's more complicated
so, assign-deep *does* deep assign values, but it also modifies the
references of the input objects.

to explain this, here's a first example

```js
> a = { foo: 3 }
{ foo: 3 }
> b = { foo: 9 }
{ foo: 9 }
> assign({}, a, b)
{ foo: 9 }
> a
{ foo: 3 }
> b
{ foo: 9 }
>
```

this is what I had expected. the `assign()` call copies foo=9 over
to the new object without modifying `a`.

but now let's look at a second example:

```js
> a = { foo: { bar: 3 } }
{ foo: { bar: 3 } }
> b = { foo: { bar: 9 } }
{ foo: { bar: 9 } }
> assign({}, a, b)
{ foo: { bar: 9 } }
> a
{ foo: { bar: 9 } }
> b
{ foo: { bar: 9 } }
```

as we see, in this case `assign()` had also modified the original
value of `a`.

this is what broke my test, as this caused `getAllMergedThemes`
to overwrite theme data from the last theme to all other themes.

I fixed it by combining `assign` with a deep `merge` to create
a deep copy of the data first.